blog edit working

+++ src/pages/BlogDetails.vue
@@ -7,12 +7,36 @@
           <p>{{blog.creator.name}}</p>
         </div>
       </div>
-      <div class="col-8 bg-light">
-        <i v-if="profile.id == blog.creator.id" class="fa fa-trash" aria-hidden="true" @click="deletePost"></i>
-        <div class="h1">{{blog.title}}</div>
+      <div v-if="isedit == false" class="col-8 bg-light">
+        <div class="row justify-content-end">
+        <div class="card col-1 bg-dark text-light m-1" v-if="profile.id == blog.creator.id">
+          <h4 class="text-center">
+            <i class="m-1 fa fa-pen-square" aria-hidden="true" @click="toggle"></i>
+            <i  class="m-1 fa fa-trash" aria-hidden="true" @click="deletePost"></i>
+          </h4>
+        </div>
+        </div>
+        <h1>{{blog.title}}</h1>
         <div class="card">
           <p>{{blog.body}}</p>
         </div>
+      </div>
+      <div v-else class="col-8 bg-light">
+        <div class="row justify-content-end">
+        <div class="card col-1 bg-dark text-light m-1" v-if="profile.id == blog.creator.id">
+          <h4 class="text-center">
+            <i class="m-1 fa fa-pen-square" aria-hidden="true" @click="toggle"></i>
+            <i  class="m-1 fa fa-trash" aria-hidden="true" @click="deletePost"></i>
+          </h4>
+        </div>
+        </div>
+        <form @submit.prevent="editblog">
+        <input type="text" name="title" class="form-control" :placeholder="blog.title" v-model="blogData.title"></input>
+        <div class="card">
+          <textarea type="text" name="body" class="form-control" :placeholder="blog.body" v-model="blogData.body"></textarea>
+          <button type="submit" class="btn btn-outline-primary">Submit Edit</button>
+        </div>
+        </form>
       </div>
     </div>
     <div class="row">
@@ -45,7 +69,9 @@ export default {
   name: "BlogDetails",
   data(){
     return{
-      newCom:{}
+      newCom:{},
+      blogData: {},
+      isedit: false
     }
   },
   computed:{
@@ -57,7 +83,7 @@ export default {
     },
     profile (){
       return this.$store.state.profile
-    }
+    },
   },
     mounted() {
     this.$store.dispatch("getBlogById", this.$route.params.id)
@@ -69,7 +95,21 @@ export default {
       this.$store.dispatch("createCom", this.newCom)
     },
     deletePost(){
-      this.$store.dispatch("deletePost", this.$route.params.id)
+      if(window.confirm("Delete for always?")){
+        this.$store.dispatch("deletePost", this.$route.params.id)
+      }else{
+        return
+      }
+    },
+    toggle(){
+      this.isedit = !this.isedit
+    },
+    editblog(){
+      this.blogData.id = this.$route.params.id
+      this.$store.dispatch("editBlog", this.blogData)
+      this.toggle()
+      this.$store.dispatch("getBlogById", this.$route.params.id
+      )
     }
   },
   components:{
